==> 04-Calculate_2Ddescriptors.py <==
from rdkit import Chem
from rdkit import ML
from rdkit.ML import Descriptors
from rdkit.ML.Descriptors import MoleculeDescriptors
import numpy as np
import pandas as pd

def ReadMoleculesSMILES(smilesfile):
    with open(smilesfile, 'r') as f:
        molList = []
        for line in f:
            name = line.split("\t")[1].strip()
            try:
                smiles = Chem.MolFromSmiles(line)
                smiles.SetProp("_Name",  name)
                molList.append(smiles)
            except:
                print(name,  "cannot be treated with RDKit")
        print("Molecules:",  len(molList))
        return molList
        
def GenerateDescriptors(molList,  descriptorsList,  label):
        print("Descriptors:",  len(descriptorsList))
        calculator = MoleculeDescriptors.MolecularDescriptorCalculator(descriptorsList)
        print(calculator.GetDescriptorNames()) #Description of the Molecular Descriptors
        
        baseData= np.arange(1, len(descriptorsList)+3)
        for mol in molList:
            descriptors = list(calculator.CalcDescriptors(mol))
            descriptors.insert(0, mol.GetProp("_Name"))
            descriptors.insert(1, label)
            baseData=np.vstack([baseData, descriptors])
        descriptors = pd.DataFrame(data=baseData[1:],columns=["Name"] + ["Group"] +descriptorsList)
        return descriptors

###200 descriptors from RDKit
descriptors2D = ["MaxEStateIndex", "MinEStateIndex", "MaxAbsEStateIndex", "MinAbsEStateIndex", "EState_VSA1", "EState_VSA2", "EState_VSA3", 
"EState_VSA4", "EState_VSA5", "EState_VSA6", "EState_VSA7", "EState_VSA8", "EState_VSA9", "EState_VSA10", "EState_VSA11", "VSA_EState1",  
"VSA_EState2", "VSA_EState3", "VSA_EState4", "VSA_EState5",  "VSA_EState6", "VSA_EState7", "VSA_EState8", "VSA_EState9", "VSA_EState10",
"MaxPartialCharge", "MinPartialCharge", "MaxAbsPartialCharge", "MinAbsPartialCharge",  "PEOE_VSA1", "PEOE_VSA2", "PEOE_VSA3", "PEOE_VSA4", 
"PEOE_VSA5", "PEOE_VSA6", "PEOE_VSA7", "PEOE_VSA8", "PEOE_VSA9", "PEOE_VSA10",  "PEOE_VSA11", "PEOE_VSA12", "PEOE_VSA13", "PEOE_VSA14", 
"FpDensityMorgan1", "FpDensityMorgan2", "FpDensityMorgan3", "LabuteASA","BalabanJ", "BertzCT", "Chi0", "Chi0n", "Chi0v", "Chi1", "Chi1n", 
"Chi1v", "Chi2n", "Chi2v", "Chi3n", "Chi3v", "Chi4n", "Chi4v", "HallKierAlpha", "Ipc", "Kappa1", "Kappa2", "Kappa3", 
"SMR_VSA1", "SMR_VSA2", "SMR_VSA3", "SMR_VSA4", "SMR_VSA5", "SMR_VSA6", "SMR_VSA7", "SMR_VSA8", "SMR_VSA9", "SMR_VSA10",
"SlogP_VSA1",  "SlogP_VSA2", "SlogP_VSA3", "SlogP_VSA4", "SlogP_VSA5", "SlogP_VSA6", "SlogP_VSA7", "SlogP_VSA8", "SlogP_VSA9", "SlogP_VSA10", 
"SlogP_VSA11", "SlogP_VSA12","qed", "MolWt", "HeavyAtomMolWt", "ExactMolWt", "NumValenceElectrons", "NumRadicalElectrons", "TPSA",  "FractionCSP3",
"HeavyAtomCount", "NHOHCount", "NOCount", "NumAliphaticCarbocycles", "NumAliphaticHeterocycles", "NumAliphaticRings", "NumAromaticCarbocycles", 
"NumAromaticHeterocycles", "NumAromaticRings", "NumHAcceptors", "NumHDonors", "NumHeteroatoms", "NumRotatableBonds", "NumSaturatedCarbocycles", 
"NumSaturatedHeterocycles", "NumSaturatedRings", "RingCount", "MolLogP", "MolMR", 
"fr_Al_COO",  "fr_Al_OH",  "fr_Al_OH_noTert", "fr_ArN",  "fr_Ar_COO",  "fr_Ar_N",  "fr_Ar_NH",  "fr_Ar_OH",  "fr_COO",  "fr_COO2",  "fr_C_O",  "fr_C_O_noCOO", "fr_C_S", 
"fr_HOCCN",  "fr_Imine",  "fr_NH0",  "fr_NH1",  "fr_NH2",  "fr_N_O",  "fr_Ndealkylation1",  "fr_Ndealkylation2",  "fr_Nhpyrrole",  "fr_SH",  "fr_aldehyde",  "fr_alkyl_carbamate",  
"fr_alkyl_halide",  "fr_allylic_oxid",  "fr_amide",  "fr_amidine",  "fr_aniline",  "fr_aryl_methyl",  "fr_azide",  "fr_azo",  "fr_barbitur",  "fr_benzene",  "fr_benzodiazepine", 
"fr_bicyclic",  "fr_diazo",  "fr_dihydropyridine",  "fr_epoxide",  "fr_ester",  "fr_ether",  "fr_furan",  "fr_guanido",  "fr_halogen",  "fr_hdrzine",  "fr_hdrzone", 
"fr_imidazole",  "fr_imide",  "fr_isocyan",  "fr_isothiocyan",  "fr_ketone",  "fr_ketone_Topliss", "fr_lactam",  "fr_lactone",  "fr_methoxy",  "fr_morpholine", 
"fr_nitrile",  "fr_nitro",  "fr_nitro_arom",  "fr_nitro_arom_nonortho",  "fr_nitroso",  "fr_oxazole",  "fr_oxime",  "fr_term_acetylene",  "fr_para_hydroxylation", 
"fr_phenol",  "fr_phenol_noOrthoHbond",  "fr_phos_acid",  "fr_phos_ester",  "fr_piperdine", "fr_piperzine",  "fr_priamide",  "fr_prisulfonamd",  "fr_pyridine",  
"fr_quatN",  "fr_sulfide",  "fr_sulfonamd",  "fr_sulfone",  "fr_tetrazole",  "fr_thiazole",  "fr_thiocyan",  "fr_thiophene",  "fr_unbrch_alkane",  "fr_urea"]

activesfile = "/home/mjose/Documentos/UOC_Master_Bioestadistica/10-TFM/03-DataSets/actives_DPPIV_std.smi" 
descriptors = GenerateDescriptors(ReadMoleculesSMILES(activesfile), descriptors2D,  1)
descriptors.to_csv('/home/mjose/Documentos/UOC_Master_Bioestadistica/10-TFM/04-Descriptors/actives_DPPIV_Descriptors2D.csv', sep= ",",  index=False, header=True) 

inactivesfile = "/home/mjose/Documentos/UOC_Master_Bioestadistica/10-TFM/03-DataSets/inactives_DPPIV_std.smi" 
descriptors = GenerateDescriptors(ReadMoleculesSMILES(inactivesfile), descriptors2D,  0)
descriptors.to_csv('/home/mjose/Documentos/UOC_Master_Bioestadistica/10-TFM/04-Descriptors/inactives_DPPIV_Descriptors2D.csv', sep= ",",  index=False, header=True)
